fix building with RPi support
When building an executable which requires gocec which in turn builds on
top of libcec compiled with Raspberry Pi support a build failure occurs
because some of the libraries required by libcec can't be found and
specifying the -rpath is required.
This change doesn't hurt building on systems without this directory so
it can always be included without a problem.