clear configuration before using it
The CEC config must be cleared before using it. This as clearing it
makes sure all the proper values/states are set. This as for example the
devices list needs to be initialized with all values being
`CEC_DEVICE_TYPE_RESERVED` (2) while CGO initializes all values with 0
which corresponds to `CEC_DEVICE_TYPE_TV`. And most likely there are
some other quirks because of wrongly initialized values.